4.20-4.23

+++ controllers/blogRouter.js
@@ -3,28 +3,24 @@ const jwt = require("jsonwebtoken");
 const Blog = require("../models/blog");
 const User = require("../models/user");
 
-const getTokenFrom = request => {
-	const authorization = request.get('authorization')
-	if (authorization && authorization.startsWith('Bearer ')) {
-	  return authorization.replace('Bearer ', '')
-	}
-	return null
-  }
-
 blogRouter.get("/", async (request, response) => {
-	const blogs = await Blog.find({}).populate('user')
+	const blogs = await Blog.find({}).populate("user");
+	console.log("Get blogs:", blogs)
 	response.json(blogs);
 });
 
 blogRouter.post("/", async (request, response) => {
-	console.log("Request to create blog received.")
-	const decodedToken = jwt.verify(getTokenFrom(request), process.env.LOGIN_SECRET)
-	console.log("Decoded token:", decodedToken)
-	if (!decodedToken.id) {
-		return response.status(401).json({ error: 'token invalid' })
+	console.log("Request to create blog received.");
+	if (!request.user) {
+		return response.status(401).json({ error: "user login required" })
 	}
+	// const decodedToken = jwt.verify(request.token, process.env.LOGIN_SECRET);
+	// console.log("Decoded token:", decodedToken);
+	// if (!decodedToken.id) {
+	// 	return response.status(401).json({ error: "token invalid" });
+	// }
 
-	const user = await User.findById(decodedToken.id)
+	const user = await User.findById(request.user.id);
 
 	const blog = new Blog({
 		...request.body,
@@ -34,19 +30,35 @@ blogRouter.post("/", async (request, response) => {
 	console.log("Creating:", blog, "User: ", user);
 	const blogCreated = await blog.save();
 
-	user.blogs = user.blogs.concat(blogCreated._id)
-	await user.save()
+	user.blogs = user.blogs.concat(blogCreated._id);
+	await user.save();
 
 	response.status(201).json(blogCreated);
 });
 
-blogRouter.delete("/:id", (request, response) => {
+blogRouter.delete("/:id", async (request, response) => {
 	const id = request.params.id;
-	Blog.findByIdAndDelete(id)
-		.then((result) => response.status(204).end())
-		.catch((error) => {
-			response.status(400).json({ error: error.message });
-		});
+	console.log("Request to delete blog received:", id);
+
+	if (!request.user) {
+		return response.status(401).json({ error: "user login required" })
+	}
+	// const decodedToken = jwt.verify(request.token, process.env.LOGIN_SECRET);
+	// console.log("Decoded token:", decodedToken);
+	// if (!decodedToken.id) {
+	// 	return response.status(401).json({ error: "token invalid" });
+	// }
+
+	const blogToDelete = await Blog.findById(id).populate("user");
+	console.log("blogToDelete", blogToDelete)
+
+	if (blogToDelete.user.id!=request.user.id) {
+		return response.status(401).json({ error: "blog to delete is not created by you" });
+	}
+
+	await Blog.findByIdAndDelete(id)
+
+	response.status(204).end()
 });
 
 blogRouter.put("/:id", (request, response) => {
@@ -55,6 +67,7 @@ blogRouter.put("/:id", (request, response) => {
 			response.json(result);
 		})
 		.catch((error) => {
+			console.log("Update error:", error.message)
 			response.status(400).json({ error: error.message });
 		});
 });
